Fix bug caused by older commit
Changing User constructor parameters caused a breaking change
which needed a lot of changing. So, this commit rolls back the
last commit. Instead, to keep it simple, in a single line
in Client.js that is calling User constructor, client is removed
from its arguments keeping arguments list as expected from
the constructor.

Apparently, in Guild structure, the Member constructor was
also being wrongly called with client as an extra parameter.
This commit fixes it.